refactor display

+++ i3ipc.h
@@ -6,7 +6,7 @@ typedef struct
   unsigned char num;
   unsigned char visible;
   unsigned char urgent;
-  unsigned int x;
+  short int x;
 } Workspace;
 typedef struct
 {
